refactor(bot): break trivial circular deps (Cluster A partial + D)
PR 1 of 4 from the /refactor-pipeline plan for #871 (see
docs/decisions/2026-05-16-next-refactor-target-bot-circular-deps.md).

Madge baseline: 15 cycles. After this PR: 12 cycles.

Fixes:

Cluster D — dead self-cycle barrels (cycles 14, 15):
  - packages/bot/src/functions/download/utils/downloadVideo.ts
  - packages/bot/src/functions/download/utils/ytDlpUtils/downloader.ts

Both were one-liners: `export type * from './<self>'`. With both
`./<name>.ts` AND `./<name>/` present, TypeScript resolved to the file,
producing an infinite self-import. Zero external callers; pure dead code.
Deleted.

Cluster A (partial) — extract CustomClient from the types barrel:
  - New file packages/bot/src/types/CustomClient.ts holds the
    `CustomClient` + `CommandType` declarations (previously inline in
    types/index.ts).
  - types/index.ts becomes a pure re-export barrel (no declarations).
  - types/CommandData.ts and types/QueueMetadata.ts import `CustomClient`
    from `./CustomClient` directly, breaking the old back-edge through
    `./index`.

Fixes cycle 2 (`types/index.ts → QueueMetadata → types/index.ts`) and
the runtime barrel cycle. Cycle 1 (the type-only
`CustomClient → models/Command → types/CommandData` path) persists but
is `import type` only — erased by tsc, no runtime effect. Deferred to a
follow-up after Cluster C lands because the structural-type break
cascades into `help.ts` and similar `Collection<string, Command>`
consumers that expect the real class; cleaner to revisit alongside the
command-loader contracts.

No caller changes required — all existing `import { CustomClient } from '.../types'`
paths still resolve through the barrel re-export.

Validation:
  - npx madge --circular packages/bot/src → 12 (was 15)
  - npx tsc --noEmit clean (no new errors)

Next: PR 2 will tackle Cluster B (monitoring telemetry cycle).
