--- packages/bot/src/functions/download/utils/downloadVideo.ts
@@ -1,1 +1,0 @@
-export type * from './downloadVideo'
