--- packages/bot/src/functions/download/utils/ytDlpUtils/downloader.ts
@@ -1,1 +1,0 @@
-export type * from './downloader'
